Add proper event function to each button
- Each button now will call the game function and input the right
parameters to the game function. the round results are displayed
in the console for now.
- A very interesting problem in this section! I used the textcontent of
the buttons for the playerSelection input. I had a very hard time
printing the results of the game using a console.log.
- It turned out
that the textContent of the buttons was inputting this in the
playerSelection variable each time: "/n        rock/n        "
so I had to hack around trimming the playerSelection in order to get
things to work.